moved getHomePage and RolePage
