+++ frontend/flutter/lib/core/auth/presentation/signup_page.dart
@@ -2,12 +2,12 @@ import 'package:fresh_nest/core/auth/application/auth.dart';
 import 'package:fresh_nest/core/auth/presentation/login_page.dart';
 import 'package:fresh_nest/core/home/presentation/widgets/loading_widget.dart';
 import 'package:fresh_nest/globals.dart';
-import 'package:fresh_nest/main.dart';
 import 'package:fresh_nest/widgets/action_button.dart';
 import 'package:fresh_nest/widgets/custom_text_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+import 'package:fresh_nest/splash.dart';
 
 class SignUpPage extends ConsumerStatefulWidget {
   const SignUpPage({super.key});
